chore: rename dirs of form elements for trading
