+++ libs/ui-toolkit/src/components/index.ts
@@ -6,18 +6,18 @@ export * from './background-video';
 export * from './breadcrumbs';
 export * from './button';
 export * from './callout';
-export * from './checkbox';
+export * from './trading-checkbox';
 export * from './copy-with-tooltip';
 export * from './dialog';
 export * from './divider';
 export * from './drawer';
 export * from './dropdown-menu';
-export * from './form-group';
+export * from './trading-form-group';
 export * from './healthbar';
 export * from './icon';
 export * from './indicator';
-export * from './input-error';
-export * from './input';
+export * from './trading-input-error';
+export * from './trading-input';
 export * from './key-value-table';
 export * from './link';
 export * from './loader';
