+++ libs/ui-toolkit/src/components/trading-checkbox/checkbox.spec.tsx
