+++ libs/ui-toolkit/src/components/trading-checkbox/checkbox.stories.tsx
