+++ libs/ui-toolkit/src/components/trading-form-group/form-group.spec.tsx
