+++ libs/ui-toolkit/src/components/trading-input/index.ts
