+++ libs/ui-toolkit/src/components/trading-input/input.spec.tsx
