+++ libs/ui-toolkit/src/components/trading-input/input.stories.tsx
