+++ libs/ui-toolkit/src/components/trading-select/select.spec.tsx
