+++ libs/ui-toolkit/src/components/trading-text-area/index.ts
