+++ libs/ui-toolkit/src/components/trading-text-area/text-area.spec.tsx
