+++ libs/ui-toolkit/src/components/trading-text-area/text-area.tsx
